Generate the authentication.
Create a database seed that inserts 50 users

Create a new model called car and add 3 properties to it: Make, Model, and year

Create a factory for cars that sets the make field to either ford / honda / toyota

Create a database seed to generate 50 cars.

Create a controller, route, and view to show a contact form and add it to the main menu of your site.

Create a controller, route, and view to show an about page and add it to the main menu of your site.